move wrapSVGText to its own file
