+++ src/components/visualizations/multiline.js
@@ -5,6 +5,7 @@ import {
 	curveCatmullRom,
 	extent,
 	line,
+	merge,
 	nest,
 	scaleLinear,
 	scaleTime,
@@ -12,47 +13,10 @@ import {
 	selectAll,
 	voronoi
 } from 'd3';
-import { merge } from 'd3-array';
+import wrapSVGText from '../../utils/wrapSVGText';
 import styled, { withTheme } from 'styled-components';
 
 import DropdownSelect from '../dropdownSelect';
-
-function wrapSVGText(text, width){
-	text.each(function(){
-		let text = d3Select(this),
-			words = text.text().split(/\s+/).reverse(),
-			word,
-			line = [],
-			lineNumber = 0,
-			lineHeight = 1, // ems
-			x = text.attr('x'),
-			y = text.attr('y'),
-			dy = 0.4,
-			tspan = text
-				.text(null)
-				.append('tspan')
-				.attr('x', x)
-				.attr('y', y)
-				.attr('dy', dy + 'em');
-		while ((word = words.pop())) {
-			line.push(word);
-			tspan.text(line.join(' '));
-			if (tspan.node().getComputedTextLength() > width) {
-				line.pop();
-				tspan.text(line.join(' '));
-				line = [
-					word
-				];
-				tspan = text
-					.append('tspan')
-					.attr('x', x)
-					.attr('y', lineNumber + y)
-					.attr('dy', ++lineNumber * lineHeight + dy + 'em')
-					.text(word);
-			}
-		}
-	});
-}
 
 const FullWidthDiv = styled.div`
 	width: 100vw;
